fix: rename files form store to merchant
