+++ apps/gauzy/src/app/pages/inventory/components/manage-merchants/merchant-form/merchant-form.component.html
@@ -17,7 +17,9 @@
 
 	<nb-card-body>
 		<form [formGroup]="form" *ngIf="form">
-			<nb-stepper #stepper [disableStepNavigation]="!merchant">
+			<!-- //tstodo -->
+			<!-- <nb-stepper #stepper [disableStepNavigation]="!merchant"> -->
+				<nb-stepper #stepper >
 				<nb-step label="Main">
 					<div class="row">
 						<div class="col-md-3 col-sm-12 image-container">
