+++ apps/gauzy/src/app/pages/inventory/inventory-routing.module.ts
@@ -159,7 +159,7 @@ const routes: Routes = [
 		]
 	},
 	{
-		path: 'stores',
+		path: 'merchants',
 		component: MerchantComponent,
 		canActivate: [NgxPermissionsGuard],
 		data: {
