+++ packages/core/src/app.module.ts
@@ -129,7 +129,7 @@ import { resolveServeStaticPath } from './helper';
 import { AccountingTemplateModule } from './accounting-template/accounting-template.module';
 import { SeederModule } from './core/seeds/seeder.module';
 import { WarehouseModule } from './warehouse/warehouse.module';
-import { MerchantModule } from 'product-store/product-store.module';
+import { MerchantModule } from 'merchant/merchant.module';
 
 const { unleashConfig } = environment;
 if (unleashConfig.url) {
